# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game API",
        "version": "0.1"
    },
    "paths": {
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get a list of games",
                "description": "Returns a paginated list of games.",
                "operationId": "getGameList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Game"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Meta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Create a new game",
                "description": "Create a new game with the provided name.",
                "operationId": "createGame",
                "requestBody": {
                    "description": "Game name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the game",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get game by ID",
                "description": "Returns a single game",
                "operationId": "getResourceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Game not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Games"
                ],
                "summary": "Update a game",
                "description": "Update an existing game by ID.",
                "operationId": "updateGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Game name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the game",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Delete a game",
                "description": "Delete an existing game by ID.",
                "operationId": "deleteGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the game to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Game deleted successfully"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Game": {
                "title": "Game",
                "description": "Game object",
                "properties": {
                    "id": {
                        "description": "ID of the game",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the game",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID of the user who created the game",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Date and time when the game was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the game was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Meta": {
                "title": "Meta",
                "description": "Meta information for pagination",
                "properties": {
                    "page": {
                        "description": "page number",
                        "type": "integer"
                    },
                    "per_page": {
                        "description": "Items per page",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}